requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
googlemaps==4.4.5
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
requests==2.25.1
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
pyyaml==5.4.1